refactor(circuit): split discretized-phi operators out of circuit_utils [Clean Code: SRP]
`circuit_utils.py` is a junk drawer of ~11 unrelated responsibilities;
the refactoring plan calls for splitting it along those groups (Batch
B1). First slice: the discretized-phi basis operators.

Move to a new `scqubits/core/discretized_phi_operators.py`:
- `_diag_from_function`
- `_identity_phi`, `_phi_operator`
- `_i_d_dphi_operator`, `_i_d2_dphi2_operator`
- `_cos_phi`, `_sin_phi`

`circuit_utils.py` re-exports the symbols (the symbols are private but
imported across modules — preserving the import surface keeps the move
internal). No call site needs to change.

Verified: full pytest + mypy clean; subprocess import-order test still
passes with the new module added to the rotation.
